Python file + Instructions
Mike's special words is a string in which when split in the middle, gives two halfs of the same characters with the same frequency. When given an odd length string the middle character is ignored. Your job is to write a program to print out if something is in fact a special word.

Input Format

Input: The first line of the input contains a single integer Q, (1<Q<=100) which is the number of data sets that follow. Each data set will have string of characters.

Output Format

Output: For each data set, should generate "YES" if it is a special word or "NO" if it is not.

Sample Input

Input:

5
yoyo
tutor
lollol
jijij
topic

Sample Output

Output:

YES
NO
YES
YES
NO

Explanation

Yoyo when broke up is "yo" and "yo". While tutor when broken is "tu" and "or"